requirements.txt
amqp==2.5.2
asn1crypto==0.24.0
Babel==2.6.0
billiard==3.6.3.0
celery==4.4.2
cffi==1.12.2
Click==7.0
cryptography==2.6.1
decorator==4.4.0
Flask==1.1.2
Flask-Caching==1.8.0
Flask-Cors==3.0.8
Flask-Headers==1.0
flask-mongoengine==0.9.5
Flask-Redis==0.4.0
Flask-Script==2.0.6
Flask-WTF==0.14.3
gevent==20.5.0
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
mongoengine==0.17.0
numpy==1.16.2
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.7.2
pytz==2018.9
raven==6.10.0
redis==3.2.1
six==1.12.0
supervisor==4.2.0
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.2.1
uwsgi==2.0.18